requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.3
PyYAML==6.0.2
requests==2.32.3
yattag==1.16.1
